requirements.txt
# Medical Red-Teaming Framework Requirements
# Core data processing
pandas>=1.3.0

# Data validation and settings
pydantic>=1.8.0

# AI/ML API clients
openai>=1.0.0
anthropic>=0.7.0
google-genai>=1.27.0
openai-agents>=0.2.3
openpyxl>=3.1.5

# Progress bars and utilities
tqdm>=4.60.0

# Optional: For better development experience
# jupyter>=1.0.0
# matplotlib>=3.5.0
# seaborn>=0.11.0 